register_Login

+++ BankSystemProject/Models/CreditCard.cs
@@ -5,8 +5,8 @@
         public int CreditCardId { get; set; }
         public int CustomerAccountId { get; set; }
         public string CardType { get; set; }
-        public decimal CreditLimit { get; set; }
-        public decimal Balance { get; set; }
+        public double CreditLimit { get; set; }
+        public double Balance { get; set; }
         public DateTime ExpiryDate { get; set; }
         public string Status { get; set; }
