+++ BankSystemProject/Models/Loan.cs
@@ -5,7 +5,7 @@
         public int LoanId { get; set; }
         public int CustomerAccountId { get; set; }
         public int LoanTypeId { get; set; }
-        public decimal LoanAmount { get; set; }
+        public double LoanAmount { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
         public string Status { get; set; }
