+++ BankSystemProject/Models/TransferInfo.cs
@@ -6,11 +6,11 @@
         public string UserId { get; set; }          // Foreign Key to Users
         public string AccountNumTransferFrom { get; set; }
         public string AccountNumTransferTo { get; set; }
-        public decimal BalanceFromBeforeTransfer { get; set; }
-        public decimal BalanceToBeforeTransfer { get; set; }
+        public double BalanceFromBeforeTransfer { get; set; }
+        public double BalanceToBeforeTransfer { get; set; }
         public DateTime DateTimeTransfer { get; set; }
-        public decimal BalanceFromAfterTransfer { get; set; }
-        public decimal BalanceToAfterTransfer { get; set; }
+        public double BalanceFromAfterTransfer { get; set; }
+        public double BalanceToAfterTransfer { get; set; }
 
         public Users User { get; set; }           // Navigation property to Users
     }
